MainViewModel major refactoring.
- Refactoring view to update when you change view.

- Refactored New & EditappointmentControllers

- Refreshing database with multiple clients

- Calendars class -> MainViewModel
