+++ aniwash/src/main/java/aniwash/view/ScheduleController.java
@@ -2,14 +2,17 @@ package aniwash.view;
 
 import aniwash.MainApp;
 import aniwash.resources.model.CreatePopUp;
-import aniwash.resources.model.ModelViewViewmodel;
+import aniwash.resources.model.MainViewModel;
 import com.calendarfx.view.CalendarView;
+import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.paint.Color;
 
 import java.io.IOException;
+import java.time.LocalDate;
+import java.time.LocalTime;
 
 public class ScheduleController {
     private CreatePopUp popup = new CreatePopUp();
@@ -25,11 +28,35 @@ public class ScheduleController {
     private Button listButton;
     @FXML
     private Button monthButton;
-    private ModelViewViewmodel modelViewViewmodel = new ModelViewViewmodel();
+    private MainViewModel mainViewModel = new MainViewModel();
 
     public void initialize() {
-        calendarView.getCalendarSources().addAll(modelViewViewmodel.getFamilyCalendar());
+        calendarView.getCalendarSources().addAll(mainViewModel.getFamilyCalendar());
+        calendarView.setShowDeveloperConsole(true);
+        calendarView.setRequestedTime(LocalTime.now());
         calendarView.setEntryDetailsCallback(new CreatePopUp());
+        Thread updateTimeThread = new Thread("Calendar: Update Time Thread") {
+            @Override
+            public void run() {
+                while (calendarView.isVisible()) {
+                    Platform.runLater(() -> {
+                        calendarView.setToday(LocalDate.now());
+                        calendarView.setTime(LocalTime.now());
+                        mainViewModel.updateCalendar(false);
+                    });
+                    try {
+                        // update every 10 seconds
+                        sleep(10000);
+                    } catch (InterruptedException e) {
+                        e.printStackTrace();
+                    }
+
+                }
+            }
+        };
+        updateTimeThread.setPriority(Thread.MIN_PRIORITY);
+        updateTimeThread.setDaemon(true);
+        updateTimeThread.start();
     }
 
     @FXML
@@ -53,7 +80,7 @@ public class ScheduleController {
     @FXML
     private void testWeekView() {
         calendarView.showWeekPage();
-        modelViewViewmodel.updateCalendar();
+        mainViewModel.updateCalendar(true);
         Color selected = new Color(1, 1, 1, 1).rgb(255, 255, 255);
         testbutton.setTextFill(selected);
         testbutton.styleProperty().set("-fx-background-color: #1485ff");
@@ -68,7 +95,7 @@ public class ScheduleController {
     @FXML
     private void testListView() {
         calendarView.showDayPage();
-        modelViewViewmodel.updateCalendar();
+        mainViewModel.updateCalendar(true);
         Color selected = new Color(1, 1, 1, 1).rgb(255, 255, 255);
         listButton.setTextFill(selected);
         listButton.styleProperty().set("-fx-background-color: #1485ff");
@@ -82,7 +109,7 @@ public class ScheduleController {
     @FXML
     private void testMonthlyView() {
         calendarView.showMonthPage();
-        modelViewViewmodel.updateCalendar();
+        mainViewModel.updateCalendar(true);
         Color selected = new Color(1, 1, 1, 1).rgb(255, 255, 255);
         monthButton.styleProperty().set("-fx-background-color: #1485ff");
         monthButton.setTextFill(selected);
