+++ aniwash/src/main/resources/aniwash/view/createAppoitment.fxml
@@ -14,7 +14,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="700.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="aniwash.view.NewAppoitmentController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="700.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="aniwash.view.NewAppointmentController">
     <children>
         <Circle fx:id="one" fill="DODGERBLUE" layoutX="44.0" layoutY="98.0" radius="22.0" stroke="TRANSPARENT" strokeType="INSIDE" style="-fx-fill: #d0d1d4;" AnchorPane.leftAnchor="22.0" />
         <Text fill="WHITE" layoutX="39.0" layoutY="106.0" strokeType="OUTSIDE" strokeWidth="0.0" text="1">
